requirements.txt
AllanTools==2019.9
cycler==0.10.0
kiwisolver==1.3.1
matplotlib==3.4.2
numpy==1.21.0
Pillow==8.3.1
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.4.1
scipy==1.7.0
six==1.16.0
